add tests to check for items on the dashboard
these check to see that the accessible (attr_accessible) items are shown
on the index/edit/show page for activeadmin

I added them because I forgot to show an item on the index page for one
of the models.  This can happen when you change the behavior for
activeadmin (in app/admin/model_name.rb)

You should only need to change spec/features/show_all_fields.feature
when making more/less attributes accessable

However you will need to change both the feature file and the steps file
(spec/features/step_definitions/all_fields_steps.rb) when adding a new
model.  This may seem tedious, but it ensures we show the user the right
items when changing the forms/tables in activeadmin

Note: at least one of these tests is failing now, but the tests are
correct.